[파일 삭제 및 에디터 수정] 타이틀 씬 UI 개선, 불필요 파일 정리(with 장형택)

+++ Assets/Script/Managers/GameManager.cs
@@ -15,7 +15,7 @@ public class GameManager : MonoBehaviour
 
     private bool isGameOver = false;
 
-    private double totalTime;
+    private float totalTime;
     public Text totalTimeText;
 
     private int stage;
